# ethos: move bulk of state machine out of ISR context