--- a/drivers/Makefile.dep
+++ b/drivers/Makefile.dep
@@ -48,6 +48,10 @@ ifneq (,$(filter ccs811_%,$(USEMODULE)))
   USEMODULE += ccs811
 endif
 
+ifneq (,$(filter ethos_%,$(USEMODULE)))
+  USEMODULE += ethos
+endif
+
 ifneq (,$(filter hmc5883l_%,$(USEMODULE)))
   USEMODULE += hmc5883l
 endif