--- a/drivers/ethos/Makefile
+++ b/drivers/ethos/Makefile
@@ -1,1 +1,7 @@
+# exclude submodule sources from *.c wildcard source selection
+SRC := $(filter-out stdio.c,$(wildcard *.c))
+
+# enable submodules
+SUBMODULES := 1
+
 include $(RIOTBASE)/Makefile.base