--- a/drivers/ethos/Makefile.dep
+++ b/drivers/ethos/Makefile.dep
@@ -3,3 +3,7 @@ USEMODULE += iolist
 USEMODULE += netdev_eth
 USEMODULE += random
 USEMODULE += tsrb
+
+ifneq (,$(filter ethos_stdio,$(USEMODULE)))
+  USEMODULE += isrpipe
+endif