--- a/drivers/ethos/ethos.c
+++ b/drivers/ethos/ethos.c
@@ -40,7 +40,7 @@
 #ifdef MODULE_STDIO_ETHOS
 #include "stdio_uart.h"
 #include "isrpipe.h"
-extern isrpipe_t stdio_uart_isrpipe;
+extern isrpipe_t ethos_stdio_isrpipe;
 #endif
 
 #define ENABLE_DEBUG 0
@@ -106,7 +106,7 @@ static void _handle_char(ethos_t *dev, char c)
 #ifdef MODULE_STDIO_ETHOS
         case ETHOS_FRAME_TYPE_TEXT:
             dev->framesize++;
-            isrpipe_write_one(&stdio_uart_isrpipe, c);
+            isrpipe_write_one(&ethos_stdio_isrpipe, c);
 #endif
     }
 }