--- a/drivers/ethos/stdio.c
+++ b/drivers/ethos/stdio.c
@@ -49,7 +49,33 @@ extern unsigned ethos_unstuff_readbyte(uint8_t *buf, uint8_t byte,
 
 ssize_t stdio_read(void* buffer, size_t len)
 {
-    return (ssize_t)isrpipe_read(&ethos_stdio_isrpipe, buffer, len);
+    uint8_t *ptr = buffer;
+    bool escaped = false;
+    uint8_t frametype = ETHOS_FRAME_TYPE_TEXT;
+    uint8_t byte;
+
+    do {
+        int read = isrpipe_read(&ethos_stdio_isrpipe, &byte, 1);
+        int tmp;
+
+        if (read == 0) {
+            continue;
+        }
+        else if (len == 0) {
+            return -ENOBUFS;
+        }
+        tmp = ethos_unstuff_readbyte(ptr, byte, &escaped, &frametype);
+        ptr += tmp;
+        if ((unsigned)(ptr - (uint8_t *)buffer) > len) {
+            while (byte != ETHOS_FRAME_DELIMITER) {
+                /* clear out unreceived frame */
+                isrpipe_read(&ethos_stdio_isrpipe, &byte, 1);
+            }
+            return -ENOBUFS;
+        }
+    } while (byte != ETHOS_FRAME_DELIMITER);
+
+    return ptr - (uint8_t *)buffer;
 }
 
 ssize_t stdio_write(const void* buffer, size_t len)