--- a/drivers/include/ethos.h
+++ b/drivers/include/ethos.h
@@ -34,7 +34,7 @@ extern "C" {
 #endif
 
 /* if using ethos + stdio, use STDIO_UART values unless overridden */
-#if IS_USED(MODULE_STDIO_ETHOS) || defined(DOXYGEN)
+#if IS_USED(MODULE_ETHOS_STDIO) || defined(DOXYGEN)
 #include "stdio_uart.h"
 /**
  * @defgroup drivers_ethos_config     Ethernet-over-serial driver driver compile configuration