<!--
The RIOT community cares a lot about code quality.
Therefore, before describing what your contribution is about, we would like
you to make sure that your modifications are compliant with the RIOT
coding conventions, see https://github.com/RIOT-OS/RIOT/blob/master/CODING_CONVENTIONS.md.
-->

### Contribution description
This tries to fix https://github.com/RIOT-OS/RIOT/issues/17254 by eliminating the condition that causes the race altogether. It moves the following parts of ethos' state machine out of ISR context, inspired by the way it is done in `slipdev`:

- Sending and replying to HELLO messages
- Byte-unstuffing

Some escape handling is still needed in the ISR handler, due to ethos' protocol design, to determine if a received byte must go into the netdev queue (tsrb) or the STDIO queue (isrpipe), but the actual unstuffing is now done in the STDIO and netdev handler threads, respectively.

This causes a slight increase of ROM, as byte unstuffing needs to be now handled in both stdio and the netdev context, but we also gain some RAM, as the `ethos_t` members `framesize`, `last_framesize`, and `accept_new` are not needed anymore.

We also have the added benefits, that now multiple frames can be added to the respective pipes, separated by the `FRAME_DELIMITER` byte.
<!--
Put here the description of your contribution:
- describe which part(s) of RIOT is (are) involved
- if it's a bug fix, describe the bug that it solves and how it is solved
- you can also give more information to reviewers about how to test your changes
-->


### Testing procedure
Compile and flash `examples/gnrc_border_router` to a board of your choice. To be closer to the setup in the testbed, I used an `iotlab-m3` and set `ETHOS_BAUDRATE=500000` (but I also tested with `samr21-xpro` with the same `ETHOS_BAUDRATE`):

```sh
BOARD=iotlab-m3 ETHOS_BAUDRATE=500000 make -C examples/gnrc_border_router/ flash -j term
```

get the link-layer address of the `ethos` interface (the one with 6-byte hardware address) using `ifconfig` and pound that address with multiple `ping` instances:

```sh
for _ in $(seq 5); do
    sudo ping -c 1500 -f fe80::a804:28ff:fefa:f3f8%tap0 &
done
```

On `master` I get a significantly worse result, when it comes to packet loss, as with this PR at the cost of slightly higher latency (the cost of cleaner multiplexing, I guess):

#### master
```
[…]
--- fe80::a804:28ff:fefa:f3f8%tap0 ping statistics ---
1500 packets transmitted, 890 received, 40.6667% packet loss, time 24033ms
rtt min/avg/max/mdev = 7.252/17.322/1016.708/33.617 ms, pipe 60, ipg/ewma 16.032/16.285 ms

--- fe80::a804:28ff:fefa:f3f8%tap0 ping statistics ---
1500 packets transmitted, 876 received, 41.6% packet loss, time 24037ms
rtt min/avg/max/mdev = 7.929/18.462/1019.208/47.469 ms, pipe 61, ipg/ewma 16.035/16.224 ms

--- fe80::a804:28ff:fefa:f3f8%tap0 ping statistics ---
1500 packets transmitted, 858 received, 42.8% packet loss, time 24041ms
rtt min/avg/max/mdev = 8.087/19.703/1021.549/58.747 ms, pipe 61, ipg/ewma 16.038/16.559 ms

--- fe80::a804:28ff:fefa:f3f8%tap0 ping statistics ---
1500 packets transmitted, 841 received, 43.9333% packet loss, time 24051ms
rtt min/avg/max/mdev = 7.576/17.265/1016.057/34.544 ms, pipe 60
, ipg/ewma 16.044/15.698 ms
--- fe80::a804:28ff:fefa:f3f8%tap0 ping statistics ---
1500 packets transmitted, 856 received, 42.9333% packet loss, time 24053ms
rtt min/avg/max/mdev = 8.167/17.401/1016.099/34.267 ms, pipe 60, ipg/ewma 16.045/16.766 ms
```

#### This PR
```
--- fe80::a804:28ff:fefa:f3f8%tap0 ping statistics ---
1500 packets transmitted, 1326 received, 11.6% packet loss, time 21439ms
rtt min/avg/max/mdev = 13.336/41.125/55.888/3.944 ms, pipe 4, ipg/ewma 14.301/40.972 ms
...
--- fe80::a804:28ff:fefa:f3f8%tap0 ping statistics ---
1500 packets transmitted, 1315 received, 12.3333% packet loss, time 21521ms
rtt min/avg/max/mdev = 20.581/41.220/55.949/4.130 ms, pipe 5, ipg/ewma 14.356/41.088 ms
   
--- fe80::a804:28ff:fefa:f3f8%tap0 ping statistics ---
1500 packets transmitted, 1297 received, 13.5333% packet loss, time 21639ms
rtt min/avg/max/mdev = 12.583/41.019/56.037/4.523 ms, pipe 4, ipg/ewma 14.435/24.055 ms

--- fe80::a804:28ff:fefa:f3f8%tap0 ping statistics ---
1500 packets transmitted, 1285 received, 14.3333% packet loss, time 21690ms
rtt min/avg/max/mdev = 13.235/41.065/55.867/4.555 ms, pipe 4, ipg/ewma 14.469/23.689 ms

--- fe80::a804:28ff:fefa:f3f8%tap0 ping statistics ---
1500 packets transmitted, 1287 received, 14.2% packet loss, time 21706ms
rtt min/avg/max/mdev = 12.578/41.029/56.543/4.845 ms, pipe 4, ipg/ewma 14.480/21.693 ms
```


<!--
Details steps to test your contribution:
- which test/example to compile for which board and is there a 'test' command
- how to know that it was not working/available in master
- the expected success test output
-->


### Issues/PRs references
Fixes https://github.com/RIOT-OS/RIOT/issues/17254
<!--
Examples: Fixes #1234. See also #5678. Depends on PR #9876.

Please use keywords (e.g., fixes, resolve) with the links to the issues you
resolved, this way they will be automatically closed when your pull request
is merged. See https://help.github.com/articles/closing-issues-using-keywords/.
-->