@miri64 2021-11-24
Forgot to mention: As the stdio now also needs special handling in the thread layer, this moves the `stdio_ethos` implementation to its own `ethos` submodule (maybe also a step towards https://github.com/RIOT-OS/RIOT/issues/13469).