@chrysn 2021-11-24
I'm having a look, but I'm not too familiar with the ethos subsystem overall.

If we already had subsystem maintainer reviews, would you consider yourself suitable as an ethos subsystem maintainer? If so, I think I can provide sufficient review, otherwise I'm still doing the same but would ask around for someone who can give this a high-level Go just from the description you provide in the PR.