@miri64 2021-11-24
> * In stdio_uart, `#ifdef M_S_ETHOS` / `uart_init(ETHOS_UART, ...);` remains; is that intentional? (Seems like something was missed in the split).

Arghs forgot to remove that change. Will do ASAP.