@miri64 2021-11-24
> Not related to concrete code changes (at least such that GitHub would let me annotate):
> 
>     * In stdio_uart, `#ifdef M_S_ETHOS` / `uart_init(ETHOS_UART, ...);` remains; is that intentional? (Seems like something was missed in the split).
> 
> 
> Other than that (and barring testing which I'd do on the complete thing), the ethos move LGTM.

Addressed with the other issues pointed out above.