@fjmolinas 2022-03-30
With this PR for some reason, `ethos` + `riotboot` causes the shell to continuously exit, I'm investigating but if someone has an idea:

```
FEATURES_REQUIRED=riotboot USEMODULE=riotboot BOARD=dwm1001 make -C examples/gnrc_border_router/ flash term -j7
NETOPT_RX_END_IRQ not implemented by driver
gnrc_uhcpc: Using 8 as border interface and 7 as wireless interface.
udp_sendto(): sendto(): Cannot assign requested address
main(): This is RIOT! (Version: 2022.01-devel-1018-g820a3-HEAD)
RIOT border router example application
All up, running the shell now
> > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > 
```