@fjmolinas 2022-03-31
> Its the second commit https://github.com/RIOT-OS/RIOT/commit/820a3976ad676aa94306299d74c4f7748be50629, but I didn't all the time too look into it as I hoped yesterday, is there anything obvious that could be at fault here @miri64?

I didn't test on `iotlab-m3` but `samr21-xpro` and `dwm1001`