@miri64 2022-03-31
> Its the second commit [820a397](https://github.com/RIOT-OS/RIOT/commit/820a3976ad676aa94306299d74c4f7748be50629), but I didn't all the time too look into it as I hoped yesterday, is there anything obvious that could be at fault here @miri64?

Not sure, but I observed the behavior you mention also sometimes on `iotlab-m3`. The cause of this is unclear. Maybe some combination of `Ctrl`+`D` or `Ctrl`+`C` isn't properly handled by the `stdio` part? Or does the node wake up like this in your case?