@fjmolinas 2022-03-31
> Not sure, but I observed the behavior you mention also sometimes on `iotlab-m3`. The cause of this is unclear. Maybe some combination of `Ctrl`+`D` or `Ctrl`+`C` isn't properly handled by the `stdio` part? Or does the node wake up like this in your case?

It seems to happen after flashing only, resetting the BOARD after seems to fix it though.