@miri64 2022-03-31
> It seems to happen after flashing only, resetting the BOARD after seems to fix it though.

Still not clear to me if this happens when you just started the board or if you hit any key.