@miri64 2022-03-31
Ok, that I did not observe on `iotlab-m3` :-/. Maybe there is a broken condition which makes the shell exit and then start again? Have you checked if the `ethos` part at the host side may be the cause?